+++ ros_ws/src/teleoperation/src/joystick_controller.cpp
@@ -1,16 +1,16 @@
-#include "remote_joy.h"
+#include "joystick_controller.h"
 
 /**
  * @brief Construct a new Remote Joy:: Remote Joy object
  */
-RemoteJoy::RemoteJoy()
+JoystickController::JoystickController()
 {
     this->drive_parameter_publisher =
         this->node_handle.advertise< drive_msgs::drive_param >(
             TOPIC_DRIVE_PARAMETERS, 1);
 
     this->joystick_subscriber = this->node_handle.subscribe< sensor_msgs::Joy >(
-        "joy", 10, &RemoteJoy::joystickCallback, this);
+        "joy", 10, &JoystickController::joystickCallback, this);
 }
 
 /**
@@ -21,7 +21,7 @@ RemoteJoy::RemoteJoy()
  * of
  * the buttons and axes on the gamepad
  */
-void RemoteJoy::joystickCallback(const sensor_msgs::Joy::ConstPtr& joystick)
+void JoystickController::joystickCallback(const sensor_msgs::Joy::ConstPtr& joystick)
 {
     double steering_angle = -joystick->axes[ JOYSTICK_AXIS_STEERING ];
     double velocity = (joystick->axes[ JOYSTICK_AXIS_THROTTLE ] - 1) * -0.5;
@@ -35,7 +35,7 @@ void RemoteJoy::joystickCallback(const sensor_msgs::Joy::ConstPtr& joystick)
  * @param throttle The throttle provided by the gamepad input
  * @param steering_angle The steering angle provided by the gamepad input
  */
-void RemoteJoy::publishDriveParameters(double velocity, double steering_angle)
+void JoystickController::publishDriveParameters(double velocity, double steering_angle)
 {
     drive_msgs::drive_param drive_parameters;
     drive_parameters.velocity = velocity;
@@ -46,8 +46,8 @@ void RemoteJoy::publishDriveParameters(double velocity, double steering_angle)
 
 int main(int argc, char** argv)
 {
-    ros::init(argc, argv, "remote_joystick_controller");
-    RemoteJoy remote_joy;
+    ros::init(argc, argv, "joystick_controller");
+    JoystickController joystick_controller;
 
     ros::spin();
 }
